+++ CHANGELOG.md
@@ -6,6 +6,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-09-03
+### Added
+- Geo Tools
+  - Atmospheric Utilities (AtmosUtil)
+    - Added more accurate functions to calculate various atmospheric parameters.
+- Math Tools
+  - MathUtil
+    - Added temperature and velocity conversions.
+
+### Deprecated
+- Geo Tools
+  - GeoUtil
+    - Use the function sin AtmosUtil
+      - `convertIasToTas`
+      - `convertTasToIas`
+      - `convertedPressureToDensityAlt`
+
+### Fixed
+- Magnetic Tools
+  - MagneticResult
+    - Fixed issue with fields being null
+
 ## [0.1.0] - 2021-09-02
 ### Added
 - Geo Tools
@@ -55,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Conversion Functions
   - Legendre Spherical Harmonics Functions (For Magnetic Calculations)
 
-[Unreleased]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/compare/0.1.0...HEAD
+[Unreleased]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/compare/0.2.0...HEAD
+[0.2.0]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/compare/v0.0.7...v0.1.0
 [0.0.7]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/compare/v0.0.1...v0.0.7
 [0.0.1]: https://gitlab.com/vatsim-tools/aviation-calc-util/-/releases/v0.0.1